backend_fastapi/requirements.txt
fastapi==0.115.2
pymongo==4.13.0
pydantic==2.11.4
python-multipart==0.0.20
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.1.0.post1
python-jose==3.3.0
uvicorn==0.34.2
